pyproject.toml
[tool.poetry]
name = "geneweaver-db"
version = "0.5.0a15"
description = "Database Interaction Services for GeneWeaver"
authors = ["Jax Computational Sciences <cssc@jax.org>"]
readme = "README.md"
license = "Apache-2.0"
homepage = "https://thejacksonlaboratory.github.io/geneweaver-docs/"
repository = "https://github.com/TheJacksonLaboratory/geneweaver-db"
packages = [
    { include = "geneweaver/db", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.9"
geneweaver-core = ">=0.10.0a3,<1.0.0"
psycopg = {version = "3.1.18", extras = ["binary"]}

[tool.poetry.group.dev.dependencies]
geneweaver-testing = ">=0.1.2,<1.0.0"
pytest-asyncio = "^0.23.5"

[tool.ruff]
select = ['F', 'E', 'W', 'A', 'C90', 'N', 'B', 'ANN', 'D', 'I', 'ERA', 'PD', 'NPY', 'PT']

[tool.ruff.per-file-ignores]
"tests/*" = ["ANN001", "ANN201"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
